main/serializers.py
# serializers.py
from rest_framework import serializers

from . import models


class ContactFormSerializer(serializers.Serializer):
    name = serializers.CharField(max_length=100)
    email = serializers.EmailField()
    message = serializers.CharField()


class LessonSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Lesson
        fields = ['name', 'photo', 'course_description', 'certification_info']


class CategorySerializer(serializers.ModelSerializer):
    lessons = LessonSerializer(many=True, read_only=True)

    class Meta:
        model = models.Category
        fields = ['name', 'photo', 'lessons']

